fix: define navigationListService
