+++ projects/common/src/navigation/navigation.config.ts
@@ -1,23 +1,4 @@
 import { Observable } from 'rxjs';
-import { Color } from '../color/color';
-import { FeatureState} from '../feature/state/feature.state';
-
-export type NavItemConfig = NavItemLinkConfig | NavItemHeaderConfig | NavItemDividerConfig;
-
-export interface NavItemLinkConfig {
-  type: NavItemType.Link;
-  icon: string;
-  iconSize?: string;
-  label: string;
-  matchPaths: string[]; // For now, default path is index 0
-  features?: string[];
-  replaceCurrentHistory?: boolean;
-  isBeta?: boolean;
-  trailingIcon?: string;
-  trailingIconTooltip?: string;
-  trailingIconColor?: Color;
-  featureState$?: Observable<FeatureState>;
-}
 
 export type FooterItemConfig = FooterItemLinkConfig;
