+++ projects/components/src/navigation/nav-item/nav-item.component.ts
@@ -1,7 +1,8 @@
 import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { FeatureState, NavigationParams, NavigationParamsType, NavItemLinkConfig } from '@hypertrace/common';
+import { FeatureState, NavigationParams, NavigationParamsType } from '@hypertrace/common';
 import { IconSize } from '../../icon/icon-size';
+import { NavItemLinkConfig} from '../navigation-list.service';
 
 @Component({
   selector: 'ht-nav-item',
